로그인 토큰으로 구현(미완성)

+++ src/api/book.ts
@@ -1,7 +1,7 @@
 const BASE_URL = 'http://localhost:3001'    
 
 export interface Book {
-    id: number,
+    id: string,
     color: string,
     label: string,
     author: string,
@@ -12,26 +12,41 @@ export interface Book {
     createAt?: string,
     review?: string,
     publisher?: string,
+    userId?: string
 }
 
-export const fetchBooks = async (): Promise<Book[]> => {
-    const res = await fetch(`${BASE_URL}/books`)
+export const fetchBooks = async (userId: string): Promise<Book[]> => {
+    const res = await fetch(`${BASE_URL}/books?userId=${userId}`,)
     if (!res.ok) throw new Error("데이터 불러오기 실패")
     return res.json()
 }
 
-export const createBook = async (book: Omit<Book, "id">) => {
+export const createBook = async (book: Omit<Book, "id" | "userId">, userId: string) => {
     const res = await fetch(`${BASE_URL}/books`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(book),
+        body: JSON.stringify({
+          ...book,
+          userId,
+        }),
     })
     return res.json()
 }
 
-export const updateBook = async (id: number, updatedData: Partial<Book>) => {
+export const updateBook = async (
+  id: string,
+  updatedData: Partial<Book>,
+  userId: string
+) => {
+  const check = await fetch(`${BASE_URL}/books/${id}`)
+  const book = await check.json()
+
+  if (book.userId !== userId) {
+    throw new Error('권한 없음')
+  }
+
   const res = await fetch(`${BASE_URL}/books/${id}`, {
-    method: 'PATCH', // 일부만 수정
+    method: 'PATCH',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify(updatedData),
   })
@@ -41,7 +56,14 @@ export const updateBook = async (id: number, updatedData: Partial<Book>) => {
 }
 
 
-export const deleteBook = async (id: number) => {
+export const deleteBook = async (id: string, userId: string) => {
+  const check = await fetch(`${BASE_URL}/books/${id}`)
+  const book = await check.json()
+
+  if (book.userId !== userId) {
+    throw new Error('권한 없음')
+  }
+
   await fetch(`${BASE_URL}/books/${id}`, {
     method: 'DELETE',
   })
